refactor: decouple main from orchestrators
