+++ tests/test_strategy_fallback_logging.py
@@ -13,7 +13,7 @@ from models.strategy import StrategyPlan
 def test_strategy_fallback_logs_include_reason_and_override(tmp_path):
     import types
     sys.modules['pdfkit'] = types.SimpleNamespace(configuration=lambda **kwargs: None)
-    import main
+    from logic.strategy_merger import merge_strategy_data
 
     audit = create_audit_logger("test")
     strategy = StrategyPlan.from_dict(
@@ -34,7 +34,7 @@ def test_strategy_fallback_logs_include_reason_and_override(tmp_path):
         }
     }
     classification_map = {}
-    main.merge_strategy_data(strategy, bureau_data, classification_map, audit, [])
+    merge_strategy_data(strategy, bureau_data, classification_map, audit, [])
     audit_file = audit.save(tmp_path)
     data = json.loads(audit_file.read_text())
